+++ src/Table/index.ts
@@ -1,6 +1,7 @@
 import type { App } from 'vue'
 import { config, InstallOptions } from '../utils/config'
 import ProTable from './Table.vue'
+import { ProDefineComponent } from '../types/index'
 
 ProTable.installl = (app: App, options?: InstallOptions) => {
   const _options = Object.assign({}, config, options)
@@ -10,4 +11,17 @@ ProTable.installl = (app: App, options?: InstallOptions) => {
   app.component(ProTable.name || 'ProTable', ProTable)
 }
 
-export default ProTable
+export default ProTable as ProDefineComponent<{
+  selection: boolean | Record<string, unknown>
+  expand: boolean | Record<string, unknown>
+  index: boolean | Record<string, unknown>
+  menu: boolean | Record<string, unknown>
+  columns: Record<string, unknown>[]
+  total: number
+  pageSize: number
+  currentPage: number
+  pagination?: Record<string, unknown>
+  showOverflowTooltip: boolean
+  align?: 'left' | 'center' | 'right'
+  headerAlign?: 'left' | 'center' | 'right'
+}>