+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.4.0
+
+feat
+
+- InputTag|AutocompleteTag: add `trigger` `type` `hit` `color` `effect` prop
+
+fix
+
+- `filterRouterByHidden` without check the deep children
+
+other
+
+- InputTag|AutocompleteTag: remover `tag` prop
+- add tags and attributes components for vetur
+
 ## 0.3.0
 
 feat